对接接口。

+++ app/src/main/java/com/shangyi/kt/ui/goods/bean/GoodsSpecBean.kt
@@ -13,7 +13,7 @@ data class GoodsSpecBean(
 data class SkuBean(
     val child: List<SkuBean?>?,
     val name: String,
-    val price: Long,
+    val price: Float,
     val stock: Int,
     val image: String,
     val id: Int
